requirements.txt
python-whois
requests
python-dotenv